+++ internal/plugins/plugins_test.go
@@ -125,11 +125,13 @@ func TestAdd(t *testing.T) {
 	})
 
 	t.Run("when plugin name is valid but URL is invalid prints an error", func(t *testing.T) {
+		t.Setenv("LANG", "C") // git error messages are locale dependent
+
 		conf := config.Config{DataDir: testDataDir}
 
 		err := Add(conf, "foo", "foobar", "")
 
-		assert.ErrorContains(t, err, "unable to clone plugin: repository not found")
+		assert.ErrorContains(t, err, "unable to clone plugin: fatal: repository 'foobar' does not exist")
 	})
 
 	t.Run("when plugin name and URL are valid installs plugin", func(t *testing.T) {
@@ -266,7 +268,7 @@ func TestUpdate(t *testing.T) {
 			givenName:   "badplugin",
 			givenRef:    "",
 			wantSomeRef: false,
-			wantErrMsg:  "unable to open plugin Git repository: repository does not exist",
+			wantErrMsg:  "not a git repository",
 		},
 		{
 			desc:        "updates plugin when plugin with name exists",